Move sample tfrecord generation helper functions to test_utils.py
