+++ src/detext/utils/misc_utils.py
@@ -404,54 +404,3 @@ def estimate_train_steps(input_pattern, num_epochs, batch_size, isTfrecord):
         batch_size))
 
     return num_train_steps
-
-
-def _bytes_feature(value):
-    """Returns a bytes_list feature"""
-    if isinstance(value, type(tf.constant(0))):
-        value = value.numpy()  # BytesList won't unpack a string from an EagerTensor.
-    return tf.train.Feature(bytes_list=tf.train.BytesList(value=value))
-
-
-def _float_feature(value):
-    """Returns a float_list feature"""
-    return tf.train.Feature(float_list=tf.train.FloatList(value=value))
-
-
-def _int64_feature(value):
-    """Returns an int64_list feature"""
-    return tf.train.Feature(int64_list=tf.train.Int64List(value=value))
-
-
-def create_sample_tfrecord(out_file):
-    """Creates sample tfrecord to out_file"""
-    print("Composing fake tfrecord to file {}".format(out_file))
-    with tf.io.TFRecordWriter(out_file) as writer:
-        with tf.Graph().as_default(), tf.compat.v1.Session():
-            # Example 1
-            features = {
-                "label": _float_feature(
-                    [1., 0., 0., 0.],
-                ),
-                "query": _bytes_feature([b"hello"]),
-                "wide_ftrs": _float_feature([23.0, 14.0, 44.0, -1.0, 22.0, 19.0, 22.0, 19.0]),
-                "doc_title": _bytes_feature([b"document title 1", b"title 2 ?", b"doc title 3 ?", b"doc title 4 ?"]),
-                "wide_ftrs_sp_idx": _int64_feature([1, 0, 2, 8]),
-                "wide_ftrs_sp_val": _float_feature([1.0, 0.0, 7.0, 12.0])
-            }
-            example_proto = tf.train.Example(features=tf.train.Features(feature=features))
-            writer.write(example_proto.SerializeToString())
-
-            # Example 2
-            features = {
-                "label": _float_feature(
-                    [0., 0., 1.],
-                ),
-                "query": _bytes_feature([b"hello"]),
-                "wide_ftrs": _float_feature([23.0, 14.0, 44.0, -1.0, 22.0, 19.0]),
-                "doc_title": _bytes_feature([b"document title 1 linkedin", b"doc title 2 ?", b"doc title 3 ?"]),
-                "wide_ftrs_sp_idx": _int64_feature([8, 0, 2]),
-                "wide_ftrs_sp_val": _float_feature([1.0, 0.0, 7.0])
-            }
-            example_proto = tf.train.Example(features=tf.train.Features(feature=features))
-            writer.write(example_proto.SerializeToString())
